add StreetCafeScripts object
I leveraged a PORO for the purposes of the exercise. My goal was to have an easy way to set up and tear down the database from the original csv data as I developed the required queries. I, quite unfortunately, did not commit as I was going along and you have my sincerest apologies for this giant commit and lack of insight into my process as I moved along. Effectively, I spent most of the time in my sql console until I had a viable solution, moved it to the StreetCafeScripts as a method, and ran the script to ensure appropriate behavior.