add query_notes.txt
Adds a txt file which is a collection of scratch-pad notes, partial queries, and snippets of what I worked with as I developed solutions to the problems. I thought it might help give some insight into the process I followed working solutions out. Found out the hard way that Notepad and my terminal disagree on what a single-quote is; won't be taking that strategy again in the future lol.